combine start and end dates on edit route
