+++ TeachingRecordSystem/src/TeachingRecordSystem.SupportUi/Pages/RoutesToProfessionalStatus/EditRoute/EditRouteStatusState.cs
@@ -4,6 +4,7 @@ public class EditRouteStatusState
 {
     public RouteToProfessionalStatusStatus Status { get; set; }
     public DateOnly? HoldsFrom { get; set; }
+    public DateOnly? TrainingStartDate { get; set; }
     public DateOnly? TrainingEndDate { get; set; }
     public bool? InductionExemption { get; set; }
     public bool RouteImplicitExemption { get; init; }
