--- TeachingRecordSystem/src/TeachingRecordSystem.SupportUi/Pages/RoutesToProfessionalStatus/EditRoute/StartDate.cshtml.cs
@@ -1,63 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.AspNetCore.Mvc.Filters;
-using Microsoft.AspNetCore.Mvc.RazorPages;
-
-namespace TeachingRecordSystem.SupportUi.Pages.RoutesToProfessionalStatus.EditRoute;
-
-[Journey(JourneyNames.EditRouteToProfessionalStatus), RequireJourneyInstance]
-public class StartDateModel(
-    TrsLinkGenerator linkGenerator) : PageModel
-{
-    public JourneyInstance<EditRouteState>? JourneyInstance { get; set; }
-
-    [FromQuery]
-    public bool FromCheckAnswers { get; set; }
-
-    [FromRoute]
-    public Guid QualificationId { get; set; }
-
-    public string? PersonName { get; set; }
-
-    public Guid PersonId { get; set; }
-
-    [BindProperty]
-    [DateInput(ErrorMessagePrefix = "Start date")]
-    [Required(ErrorMessage = "Enter a start date")]
-    [Display(Name = "Enter the route start date")]
-    public DateOnly? TrainingStartDate { get; set; }
-
-    public void OnGet()
-    {
-        TrainingStartDate = JourneyInstance!.State.TrainingStartDate;
-    }
-
-    public async Task<IActionResult> OnPostAsync()
-    {
-        if (!ModelState.IsValid)
-        {
-            return this.PageWithErrors();
-        }
-        await JourneyInstance!.UpdateStateAsync(state => state.TrainingStartDate = TrainingStartDate);
-        if (TrainingStartDate.HasValue && JourneyInstance!.State.TrainingEndDate is DateOnly endDate && TrainingStartDate >= endDate)
-        {
-            return Redirect(linkGenerator.RouteEditEndDate(QualificationId, JourneyInstance.InstanceId));
-        }
-        return Redirect(FromCheckAnswers ?
-            linkGenerator.RouteCheckYourAnswers(QualificationId, JourneyInstance.InstanceId) :
-            linkGenerator.RouteEditDetail(QualificationId, JourneyInstance.InstanceId));
-    }
-
-    public async Task<IActionResult> OnPostCancelAsync()
-    {
-        await JourneyInstance!.DeleteAsync();
-        return Redirect(linkGenerator.PersonQualifications(PersonId));
-    }
-
-    public override void OnPageHandlerExecuting(PageHandlerExecutingContext context)
-    {
-        var personInfo = context.HttpContext.GetCurrentPersonFeature();
-        PersonName = personInfo.Name;
-        PersonId = personInfo.PersonId;
-    }
-}
